+++ sync-client/ext.h
@@ -28,7 +28,7 @@ const string current_datetime();
 
 const string ftoa(float num);
 
-size_t pos(const string &str, const char sym);
+int pos(const string &str, const char sym);
 
 
 }
